+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Add bookmarks related actions/reducers from @twreporter/redux
 - Add bookmark functional icons from twreporter-react(main website) and export it as widget. ( The icons are fully functional )
 - Get bookmark (bookmark card in bookmark list) from @twreporter/react-components.
+- Provide widget service.
+- Provide customized widget color.
 
 
 ### 2.1.5